Use WEX and WCR compatible Airbrake version
When this gem was first put together to solve [a problem](https://github.com/DEFRA/flood-risk-engine/pull/312) in the [Flood Risk Activity service](https://github.com/DEFRA/ruby-services-team/tree/master/services/frae) it needed to stick with a version of Airbrake that was compatible to FRAE. In fact it was needing to use an older version, and the problem that caused implementing our `Airbrake.close` pattern that sparked the idea of creating this gem.

We now want to start using this gem in our other services [Waste Carriers](https://github.com/DEFRA/ruby-services-team/tree/master/services/wcr) and [Waste Exemptions](https://github.com/DEFRA/ruby-services-team/tree/master/services/wex) but they use a later 5.8 version of Airbrake.

So to support this we'll bump the version of Airbrake used in this change, as well as bumping the version of this gem to 1.0.

That way FRAE can be set to pull the 0.1.x versions of the gem, and the rest of our services can pull the 1.x versions.