Remove airbrake close hook from gem
https://eaflood.atlassian.net/browse/RUBY-1060

This is part of work to get our services using the latest version of Rails and Ruby (6 and 2.7.1).

This PR is a branch off PR #8. Rather than directly commit to @cintamani branch I've created this separate PR to highlight what I think can be dropped from the gem now to fix the issue we're seeing.

Essentially the latest version of the Airbrake gem is throwing an exception when `close()` is called, but doing it in such a way that our tests are failing, and rake tasks in apps using the gem are failing (specifically rake assets:compile).

We believe its because the gem has fixed its issues around not reporting errors raised when the app is closed, so there is no expectation that users will call close() directly.