Remove airbrake close hook from gem (#9)
https://eaflood.atlassian.net/browse/RUBY-1060

This is part of work to get our services using the latest version of Rails and Ruby (6 and 2.7.1).

Previously this gem was using version 5.8.1 and 5.3.0 of the Airbrake gem (pre-ruby upgrade for our services). Both versions define their own `at_exit` hook but don't call `Airbrake.close` in it.

We know not calling `Airbrake.close` is an issue for things like rake tasks, where if an exception is thrown the app is likely to close before Airbrake can send the notification. This is why we implement our own `at_exit hook.

This issue was highlighted in [Airbrake issue 695](https://github.com/airbrake/airbrake/issues/695), which they then went and fixed in the ***very next version*** (6.0.0)! They did this by adding a call to `Airbrake.close()` in their own `at_exit` hook.

This not only makes our `at_exit` no longer needed, it actually causes an error. If you call Airbrake.close() more than once it will throw `raise Airbrake::Error, 'this thread pool is closed already`.

This is why our tests now fail and why any app trying to use a version of this gem with a version of Airbrake greater than 5.8.1 will throw an error on exit. 😁🎉

So this change covers removing the `at_exit` hook and the call to `Airbrake.close()` from our gem.